requirements.txt
minio==7.2.12
Werkzeug==3.1.3
numpy==1.26.4
requests==2.32.3
Pillow==11.0.0
opencv-python==4.10.0.84
matplotlib==3.9.2
ultralytics==8.3.29
psycopg2-binary==2.9.10
Flask==3.0.3
Flask-WTF==1.2.2
